Wrote seqtest.py
Seqtest is a representation of the full System Model from sequence manager
to individual unit operations. The Sequence itself is a thread and each
Unit Operation is also a thread. This allows the sequence manager to send
and recieve while unit operations are being completed. This test is fully
function and is to be implemented into current System Model.